[issue-158] Starting creating commands for pulling seperate providers
