+++ cocoon/houseDatabase/mlspin_tests/test_pull_mls.py
@@ -3,7 +3,7 @@ from cocoon.houseDatabase.models import RentDatabaseModel, HomeTypeModel
 from cocoon.houseDatabase.models import MlsManagementModel
 import os
 # Import script to pull MLSPIN data
-from cocoon.houseDatabase.management.commands.mlspin.pull_mlspin import MlspinRequester
+from cocoon.houseDatabase.management.commands.pull_mlspin import MlspinRequester
 
 
 class TestPullMlspin(TestCase):
